Phase C-1 cascade: move VirtualInfo enum + PartialEq + first impl to majit-backend
VirtualInfo enum (resume.py AbstractVirtualInfo hierarchy:
VirtualObj/VStruct/VArray/VArrayStruct/VRawBuffer/VRawSlice + 6
string-and-unicode variants) plus its PartialEq impl and the first
inherent impl block (field_sources / kind / layout_summary) move to
majit-backend::resume_value.  All dependencies (ResumeValueSource,
ResumeVirtualKind, ResumeVirtualLayoutSummary, ResumeValueLayoutSummary,
DescrRef, FieldDescrInfo, ArrayDescrInfo) are already in
majit-backend or majit-ir after the previous cascade hops.

The second VirtualInfo inherent impl block (is_about_raw, allocate,
allocate_int) depends on ResumeDataDirectReader + BlackholeAllocator
(metainterp-specific) — extracted to VirtualInfoBlackholeExt trait
in metainterp.  Callers (line 6020/6045/6076 in resume.rs) reach the
methods via the same dot syntax through the trait import.

VirtualFieldSource type alias also moves; majit-metainterp re-exports
VirtualInfo + VirtualFieldSource for caller compatibility.  Dead
#[cfg(any())] placeholder blocks (317 lines) removed.
